* update asmlib • refactor types
