+++ src/ameliance-ui/components/blocks/Block.tsx
@@ -4,8 +4,8 @@ import { forwardRef } from 'react';
 import asm from 'asm-ts-scripts';
 
 import { Component } from '../_LAB/Component';
-import type { Grid } from './helpers/grid';
 import { getGridClass } from './helpers/grid';
+import type { Grid } from './types/Grid';
 
 type ComponentElementType = ComponentProps<ElementType>;
